[PL] Replace STBuilder with createST function.
