+++ ProcessLib/ProcessVariable.cpp
@@ -15,6 +15,8 @@
 #include "MeshLib/Mesh.h"
 #include "ProcessLib/BoundaryCondition/BoundaryCondition.h"
 #include "ProcessLib/BoundaryCondition/CreateBoundaryCondition.h"
+#include "ProcessLib/SourceTerms/CreateSourceTerm.h"
+#include "ProcessLib/SourceTerms/NodalSourceTerm.h"
 #include "ProcessLib/Utils/ProcessUtils.h"
 
 namespace ProcessLib
@@ -34,8 +36,7 @@ ProcessVariable::ProcessVariable(
       _initial_condition(findParameter<double>(
           //! \ogs_file_param{prj__process_variables__process_variable__initial_condition}
           config.getConfigParameter<std::string>("initial_condition"),
-          parameters, _n_components)),
-      _source_term_builder(std::make_unique<SourceTermBuilder>())
+          parameters, _n_components))
 {
     DBUG("Constructing process variable %s", _name.c_str());
 
@@ -147,8 +148,7 @@ ProcessVariable::ProcessVariable(ProcessVariable&& other)
       _shapefunction_order(other._shapefunction_order),
       _initial_condition(std::move(other._initial_condition)),
       _bc_configs(std::move(other._bc_configs)),
-      _source_term_configs(std::move(other._source_term_configs)),
-      _source_term_builder(std::move(other._source_term_builder))
+      _source_term_configs(std::move(other._source_term_configs))
 {
 }
 
@@ -201,7 +201,7 @@ ProcessVariable::createSourceTerms(
     std::vector<std::unique_ptr<NodalSourceTerm>> source_terms;
 
     for (auto& config : _source_term_configs)
-        source_terms.emplace_back(_source_term_builder->createSourceTerm(
+        source_terms.emplace_back(createSourceTerm(
             config, dof_table, _mesh, variable_id, integration_order,
             _shapefunction_order, parameters));
